Added a type cast for pandas object
Makes this Mac friendly.

+++ px4/combine_and_resample_px4_nogui.py
@@ -162,7 +162,7 @@ def combine_and_resample_px4_nogui(input_path,file_suffix=''):
         big_df['time_properformat'] = pd.to_timedelta(big_df.index/10.0**6,'S')
         
         # switch the index of the big_df to proper time delta column
-        big_df.index = pd.to_datetime(big_df.time_properformat)
+        big_df.index = pd.to_datetime(big_df.time_properformat.astype('int64'))
                
         # create the resampled  df
         resampled_df = big_df.asfreq(str(freq_arg)+freq_type,method='ffill')
